requestbuilder -> querybuilder, namespace Kurve -> __Kurve
