+++ promises.ts
@@ -1,4 +1,4 @@
-﻿namespace Kurve {
+﻿namespace __Kurve {
     
 // Adapted from the original source: https://github.com/DirtyHairy/typescript-deferred
